Proyect v2
Implementado arquitectura en User y metodo post para crear funcionando

+++ Proyect/proof/src/main/java/com/Becom/proof/Module/Models/User.java
@@ -20,7 +20,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="user_id")
-    private float userId;
+    private Long userId;
 
     @Temporal(TemporalType.DATE)
     @Column(name="user_cration_date")
